+++ app/(pages)/schedule/[day]/page.tsx
@@ -1,11 +1,9 @@
 import { fetchSchedule } from "@/lib/pretix/schedule";
-import { Select } from "components/form/select";
 import { Fullbleed } from "components/layout/fullbleed";
 import { Schedule } from "components/schedule/schedule";
 import { Title } from "components/typography/title";
-import { format, parse } from "date-fns";
 import { notFound } from "next/navigation";
-import { useCallback } from "react";
+import { SelectDay } from "./select-day";
 
 export default async function SchedulePage({
   params,
@@ -18,38 +16,12 @@ export default async function SchedulePage({
     throw notFound();
   }
 
-  // TODO: use router?
-  //   const handleDaySelected = useCallback(
-  //     (event: React.ChangeEvent<HTMLSelectElement>) => {
-  //       window.location.href = `/schedule/${event.target.value}`;
-  //     },
-  //     []
-  //   );
-
   return (
     <>
       <article className="accent-left">
         <Title highlighted>Schedule</Title>
 
-        <Select
-          id="schedule-select"
-          name="schedule-select"
-          variant="rounded"
-          //   onChange={handleDaySelected}
-          defaultValue={params.day}
-        >
-          {days.map((date) => {
-            const isoDate = format(date, "yyyy-MM-dd");
-            const dateText = format(date, "eeee, do MMMM, yyyy");
-
-            return (
-              <option key={isoDate} value={isoDate}>
-                {dateText}
-                {/* - {type} */}
-              </option>
-            );
-          })}
-        </Select>
+        <SelectDay day={params.day} days={days} />
       </article>
 
       <Fullbleed>
